Submit exercises 4.15-4.23 for part4-d (bloglist-backend)

+++ part4/bloglist-backend/controllers/blogs.js
@@ -1,8 +1,14 @@
 const blogsRouter = require('express').Router()
+const userExtractor = require('../utils/middleware').userExtractor
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const AuthenticationError = require('../utils/error').AuthenticationError
 
 blogsRouter.get('/', async (_request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
@@ -15,36 +21,84 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+// NB: only using userExtractor middleware on routes that require token authentication
+// ...else will throw errors along all routes if used on blogsRouter or app overall
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+
+  const user = await User.findById(request.user)
+  if (!user) {
+    throw new AuthenticationError('unknown user')
+  }
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes === undefined ? 0 : body.likes,
+    user: user._id
+  })
 
   const savedBlog = await blog.save()
+  await user.updateOne(
+    { $set: { blogs: user.blogs.concat(savedBlog._id) } }
+  )
+
+  // // Using save()
+  // // NB: triggers custom unique username validation error
+  // user.blogs = user.blogs.concat(savedBlog._id)
+  // await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
-  if (deletedBlog) {
-    response.status(204).end()
-  } else {
-    response.status(404).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = await User.findById(request.user)
+  if (!user) {
+    throw new AuthenticationError('unknown user')
   }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (!blogToDelete) {
+    return response.status(404).end()
+  }
+
+  if (!(blogToDelete.user.toString() === user._id.toString())) {
+    throw new AuthenticationError(
+      'you are not permitted to delete another users blog'
+    )
+  }
+
+  await Blog.findByIdAndRemove(request.params.id)
+  response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
+  const user = await User.findById(request.user)
+  if (!user) {
+    throw new AuthenticationError('unknown user')
+  }
+
   // NB: findByIdAndUpdate receives a regular JS object as parameter
   // here we are only allowing update to blog likes
   const blog = {
     likes: request.body.likes
   }
 
+  const blogToUpdate = await Blog.findById(request.params.id)
+  if (!blogToUpdate) {
+    return response.status(404).end()
+  }
+
+  if (!(blogToUpdate.user.toString() === user._id.toString())) {
+    throw new AuthenticationError(
+      'you are not permitted to modify another users blog'
+    )
+  }
+
   const updatedBlog = await Blog
     .findByIdAndUpdate(request.params.id, blog, { new: true })
-  if (updatedBlog) {
-    response.json(updatedBlog)
-  } else {
-    response.status(404).end()
-  }
+  response.json(updatedBlog)
 })
 
 module.exports = blogsRouter
